Rename Route to Slash Command
This commit renames Route and all associcated types and methods to Slash Command.
We have come to the realisation the we require a slightly different type of route
handler for things coming into the server on the /slack/interaction path.
This new Interaction Command Handler needs to sit along side Slash Command Handler
so the descision was made to rename Route to properly describe its purpose, that it
will just deal with slash commands invoked in Slack.