Respond to help (#13)
* Include InteractionType when registering a handler

Adds an InteractionType for a route, allowing for
more than one handler to be registered for a route
and the correct one be chosen based a the combination
of CallbackID and InteractionType strings.  Driven
by the fact we would like to do a different thing if
a dialog is closed vs if it is submitted

* Find handler to run for Interactions

Find the handler to run based on routes matching both the callbackID and
InteractionType.  Do some basic error handling around the JSON that is
passed to an Interaction Handler.

* Add handler to run on submission of Help Dialog

Example handler to log the information a user submits through the
HelpRequest handler

* Allow for two loggers with different signatures

Renames LogFunc to LogfFunc to signify that it implemenets Printf style
signature